backend/requirements.txt
# Financial RAG Backend Dependencies
# Python 3.11

# ============================================================================
# Web Framework
# ============================================================================
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
aiofiles==23.2.1

# ============================================================================
# Database
# ============================================================================
sqlalchemy==2.0.25
pymysql==1.1.0
cryptography==41.0.7
alembic==1.13.1
redis==5.0.1

# ============================================================================
# Vector Database (Milvus)
# ============================================================================
pymilvus==2.3.4
milvus==2.3.3

# ============================================================================
# Graph Database (Neo4j)
# ============================================================================
neo4j==5.14.1
py2neo==2021.2.4

# ============================================================================
# Object Storage (MinIO)
# ============================================================================
minio==7.2.0

# ============================================================================
# Task Queue (Celery)
# ============================================================================
celery==5.3.6
flower==2.0.1

# ============================================================================
# Machine Learning / NLP
# ============================================================================
torch==2.1.2
transformers==4.33.0
sentence-transformers==2.3.1
flagembedding==1.2.2
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10

# ============================================================================
# Document Processing
# ============================================================================
python-docx==1.1.0
python-pptx==0.6.23
PyMuPDF==1.20.1
pdfplumber==0.10.3
pandas==2.1.4
openpyxl==3.1.2
xlrd==2.0.1
PyPDF2==3.0.1

# ============================================================================
# OCR and Image Processing
# ============================================================================
pillow==10.2.0
opencv-python-headless==4.9.0.80
pytesseract==0.3.10
easyocr==1.7.1
paddleocr==2.7.0
paddlepaddle==2.6.2

# ============================================================================
# Multimodal AI
# ============================================================================
qwen-vl-utils==0.0.1
dashscope==1.14.1
zhipuai==2.1.5.20250825

# ============================================================================
# Utilities
# ============================================================================
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0
requests==2.31.0
aiohttp==3.9.1
websockets==12.0

# ============================================================================
# Data Processing
# ============================================================================
numpy==1.26.3
scipy==1.11.4
scikit-learn==1.4.0
matplotlib==3.8.2

# ============================================================================
# Logging and Monitoring
# ============================================================================
structlog==24.1.0
colorlog==6.8.0
python-json-logger==2.0.7

# ============================================================================
# Testing
# ============================================================================
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# ============================================================================
# Development Tools
# ============================================================================
black==24.1.1
flake8==7.0.0
isort==5.13.2
mypy==1.8.0

# ============================================================================
# Other
# ============================================================================
python-dateutil==2.8.2
pytz==2023.3
tenacity==8.2.3
